Implement basic admin interface with domain auth (#11)
* Add admin interface with Google Workspace auth

* feat: Add SWR and update next-auth dependency

This commit introduces the `swr` library for data fetching and caching, enhancing the application's performance. Additionally, the `next-auth` dependency has been updated to version 4.24.11, ensuring compatibility with the latest features and improvements. Changes to the `admin-registrations` component include the addition of a comment field and improved layout for the save and cancel buttons, enhancing user experience during registration edits.